+++ src/Enums/Looper.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Busarm\PhpMini\Enums;
+
+/**
+ * Event Looper types
+ * 
+ * PHP Mini Framework
+ *
+ * @copyright busarm.com
+ * @license https://github.com/Busarm/php-mini/blob/master/LICENSE (MIT License)
+ * @codeCoverageIgnore
+ */
+class Looper
+{
+    const EV        =   1;
+    const EVENT     =   2;
+    const SWOOLE    =   3;
+    const SELECT    =   4;
+}