+++ src/Interfaces/RequestInterface.php
@@ -23,6 +23,11 @@ interface RequestInterface extends ContainerInterface
      * @return string
      */
     public function correlationId();
+    
+    /**
+     * @return string
+     */
+    public function requestId();
 
     /**
      * @return string