+++ src/Service/BaseService.php
@@ -2,7 +2,7 @@
 
 namespace Busarm\PhpMini\Service;
 
-use Busarm\PhpMini\Interfaces\ServiceProviderInterface;
+use Busarm\PhpMini\Interfaces\ServiceHandlerInterface;
 use Busarm\PhpMini\Interfaces\SingletonInterface;
 use Busarm\PhpMini\Traits\Singleton;
 
@@ -13,7 +13,15 @@ use Busarm\PhpMini\Traits\Singleton;
  * @copyright busarm.com
  * @license https://github.com/Busarm/php-mini/blob/master/LICENSE (MIT License)
  */
-abstract class BaseService implements ServiceProviderInterface, SingletonInterface
+abstract class BaseService implements ServiceHandlerInterface, SingletonInterface
 {
     use Singleton;
+
+    /**
+     * Get service location for name
+     * 
+     * @param string $name
+     * @return string|null
+     */
+    abstract protected function getLocation($name);
 }